vim: ku: type file: Change cache key to be unique for each cwd
Change value of cache key to be unique for each current working directory.  The
old code uses absolute path as cache key, but it causes the problem whenever
a pattern which is the same directory with different spelling is inputted.  In
that case, cached items are mostly not matched to the input pattern.