vim: ku: Use s:api() to call source API
Add s:api() and modify to use it to call source API such as
ku#{source}#key_table().  This cahnge is just a preparation to use a metarw
scheme as a ku source.

At the first, I try to define functions such as ku#metarw_git#key_table()
on-the-fly, but Vim requires that the prefix of an autoloaded function name and
the file name which contains the function must be matched.  So this method
cannot be used.

To solve this problem, I choose to introduce s:api() and use it to abstract
calling source APIs.  The internal of APIs of a "metarw_{scheme}" source will
be handled in s:api().