vim: ku: Revise the caching of ku#available_sources() (3)
Add functions to abstract caching of special sources.  The implementation is
just a dummy to hide the detail of how special sources are cached, so the cache
is always expired and s:update_special_source_cache() will be always called for
each calling ku#available_sources().  This problem will be fixed later.