vim: ku: Refactor - for auto component completion
Revise s:text_by_auto_directory_completion() to use a:line[-1:] instead of '/'
to separate components.  Because a:line[-1:] is the separator of components in
a:line and it may be any character in near future to support to activate this
completion by a character which may or may not be '/'.