vim: ku: Add g:ku_buffer_name / Fix a bug on MS-Windows
Add g:ku_buffer_name to configure the name of the ku buffer.

The default name of the ku buffer on Microsoft Windows platforms is changed to
'[ku]' instead of '*ku'.  Because :file `='foo*'` raises E480 on Microsoft
Windows, so '*' cannot be used as a part of buffer name on the platforms.