core: Fix bugs on non-empty typeahead buffer
Ku may complete several stuffs in wrong way if the typeahead buffer is
not empty at the beginning of a ku session.  There are two bugs on this:

(a) feedkeys() is called two times - to complete the prompt and to start
    Insert mode.  Completing the prompt will be automatically invoked
    after starting Insert mode.  If there some characters at the input
    line, this completion may be behaved unexpected way.

(b) feedkeys() to start Insert mode doesn't consider about the state of
    the typeahead buffer.  If it is not empty, keys to start Insert mode
    will be feeded AFTER the keys in the typeahead buffer.  This is
    problematic; keys to start Insert mode must be feeded BEFORE them.

Fix the above bugs.