vim: ku: Refactor source repr. - "kind_names"
Source attribute "kinds" contains a list of the names of kinds.
While in the internal, variables with name "kind" hold internal
representation of kinds.  This difference may cause confusion.

So refactor internal reepresentation of source to use "kind_names"
instead of "kinds" to hold the value, and "kinds" is removed.
Note that "kinds" is still available for ku#define_source()
for users' convenience.