vim: ku: core: Allow backspacing after ACC and etc
Vim doesn't allow deleting over the start of Inserta mode if 'backspace'
doesn't have a proper value, and moving the cursor in Insert mode will
start new Insert mode.  For example, if you type
"iabc<Left>d<BS><BS>...", the first <BS> deletes "d", but the further
<BS>s don't delete anything.

In a ku session, ACC and other events are usually happened and these
features move the cursor in Insert mode, so that backspacing may not
work.

To fix this problem, set up the value of 'backspace' on start/end of
a ku session.